+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/australia/au_abn_recognizer.py
@@ -25,6 +25,8 @@ class AuAbnRecognizer(PatternRecognizer):
     This can allow a greater variety in input, for example by removing dashes or spaces.
     """
 
+    COUNTRY_CODE = "au"
+
     PATTERNS = [
         Pattern(
             "ABN (Medium)",