+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/australia/au_tfn_recognizer.py
@@ -28,6 +28,8 @@ class AuTfnRecognizer(PatternRecognizer):
     This can allow a greater variety in input, for example by removing dashes or spaces.
     """
 
+    COUNTRY_CODE = "au"
+
     PATTERNS = [
         Pattern(
             "TFN (Medium)",