+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/italy/it_identity_card_recognizer.py
@@ -24,6 +24,8 @@ class ItIdentityCardRecognizer(PatternRecognizer):
     :param supported_entity: The entity this recognizer can detect
     """
 
+    COUNTRY_CODE = "it"
+
     PATTERNS = [
         Pattern(
             "Paper-based Identity Card (very weak)",