+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/spain/es_nif_recognizer.py
@@ -16,6 +16,8 @@ class EsNifRecognizer(PatternRecognizer):
     This can allow a greater variety in input, for example by removing dashes or spaces.
     """
 
+    COUNTRY_CODE = "es"
+
     PATTERNS = [
         Pattern(
             "NIF",