+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/spain/es_passport_recognizer.py
@@ -1,11 +1,18 @@
 from typing import List, Optional
 
+import regex as re
+
 from presidio_analyzer import Pattern, PatternRecognizer
 
 
-class ItDriverLicenseRecognizer(PatternRecognizer):
+class EsPassportRecognizer(PatternRecognizer):
     """
-    Recognizes IT Driver License using regex.
+    Recognize Spanish passport numbers using regex.
+
+    Follows the format: 3 letters followed by 6 digits (e.g. AAA123456).
+
+    Reference(s):
+    https://en.wikipedia.org/wiki/Spanish_passport
 
     :param patterns: List of patterns to be used by this recognizer
     :param context: List of context words to increase confidence in detection
@@ -13,28 +20,26 @@ class ItDriverLicenseRecognizer(PatternRecognizer):
     :param supported_entity: The entity this recognizer can detect
     """
 
-    COUNTRY_CODE = "it"
+    COUNTRY_CODE = "es"
 
     PATTERNS = [
         Pattern(
-            "Driver License",
-            (
-                r"\b(?i)(([A-Z]{2}\d{7}[A-Z])"
-                r"|(U1[BCDEFGHLJKMNPRSTUWYXZ0-9]{7}[A-Z]))\b"
-            ),
-            0.2,
+            "ES_PASSPORT",
+            r"\b[A-Z]{3}[0-9]{6}\b",
+            0.05,
         ),
     ]
-    CONTEXT = ["patente", "patente di guida", "licenza", "licenza di guida"]
+
+    CONTEXT = ["pasaporte", "passport", "número de pasaporte", "passport number"]
 
     def __init__(
         self,
         patterns: Optional[List[Pattern]] = None,
         context: Optional[List[str]] = None,
-        supported_language: str = "it",
-        supported_entity: str = "IT_DRIVER_LICENSE",
+        supported_language: str = "es",
+        supported_entity: str = "ES_PASSPORT",
         name: Optional[str] = None,
-    ) -> None:
+    ):
         patterns = patterns if patterns else self.PATTERNS
         context = context if context else self.CONTEXT
         super().__init__(
@@ -43,4 +48,5 @@ class ItDriverLicenseRecognizer(PatternRecognizer):
             context=context,
             supported_language=supported_language,
             name=name,
+            global_regex_flags=re.DOTALL | re.MULTILINE | re.IGNORECASE,
         )