+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/uk/uk_nhs_recognizer.py
@@ -16,6 +16,8 @@ class NhsRecognizer(PatternRecognizer):
     This can allow a greater variety in input, for example by removing dashes or spaces.
     """
 
+    COUNTRY_CODE = "uk"
+
     PATTERNS = [
         Pattern(
             "NHS (medium)",