+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/us/medical_license_recognizer.py
@@ -18,6 +18,8 @@ class MedicalLicenseRecognizer(PatternRecognizer):
     This can allow a greater variety in input, for example by removing dashes or spaces.
     """
 
+    COUNTRY_CODE = "us"
+
     PATTERNS = [
         Pattern(
             "USA DEA Certificate Number (weak)",